Fix errors thrown when a toolbar is closed by a click event on an icon inside it.
When popup toolbars were closed in this way, the L.Marker containing the popup toolbar throws an "Uncaught TypeError: Cannot read property 'dragging' of null."

The statements causing the issue seem to be those removing event handlers in L.Toolbar.Popup#onRemove that were introduced to address #7.

Error arose unexpectedly - it may have come from merging in the fix for #3, since both that branch (single-toolbar-on-map) and the master branch contained modifications to L.Toolbar.Popup#onRemove.