Correct L.Toolbar usage in contributed subtoolbar example.
In particular, ensure that each action and sub-action is fully bound. In practice, toolbar actions should be modular - that is, developers should be able to export a toolbar action as part of a library without requiring that it be defined in the same file as the map or its parent toolbar. All values which a toolbar action needs should be passed in to its `initialize` function and stored in its context.